Add optional argument for pause time.
Importing issues or issue comments generates notification in
GitHub.  Those are subject to secondary rate limits because
they generate notifications, which in turn create may send
emails. To handle this, we introduce a pause after creation
of each issue or issue comment.

Add an optional argument for pause time.  The default pause
time if no argument is passed e is 30,000 millseconds
(30 seconds).